requirements.txt
asgiref==3.5.2
autopep8==2.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
pycodestyle==2.10.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
